+++ jadler-all/src/test/java/net/jadler/AbstractJadlerResetIntegrationTest.java
@@ -6,13 +6,12 @@ package net.jadler;
 
 import net.jadler.mocking.Verifying;
 import net.jadler.stubbing.RequestStubbing;
-import net.jadler.stubbing.server.jetty.JettyStubHttpServer;
 import org.apache.commons.httpclient.HttpClient;
 import org.apache.commons.httpclient.methods.GetMethod;
 import org.junit.After;
 import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
+
 import java.io.IOException;
 
 import static org.hamcrest.Matchers.is;
@@ -21,13 +20,10 @@ import static org.junit.Assert.assertThat;
 /**
  * Tests that its possible to reset JadlerMocker.
  */
-public class JadlerResetIntegrationTest {
-    private static final JadlerMocker mocker = new JadlerMocker(new JettyStubHttpServer());
+public abstract class AbstractJadlerResetIntegrationTest {
+    
+    protected static JadlerMocker mocker;
 
-    @BeforeClass
-    public static void start() {
-        mocker.start();
-    }
 
     @AfterClass
     public static void close() {